requirements.txt
gradio
python-dotenv
pandas
playwright
google-generativeai
requests
matplotlib
langchain_openai
scikit-learn
nltk
beautifulsoup4
bs4
browser-cookie3